move functions out of core
