+++ osmnx/plot.py
@@ -20,7 +20,7 @@ from shapely.geometry import MultiPolygon
 from . import settings
 from .core import graph_from_address
 from .core import graph_from_point
-from .core import bbox_from_point
+from .utils_geo import bbox_from_point
 from .projection import project_graph
 from .utils_graph import graph_to_gdfs
 from .simplify import simplify_graph
