+++ osmnx/save_load.py
@@ -6,7 +6,6 @@
 ################################################################################
 
 import ast
-import bz2
 import geopandas as gpd
 import os
 import networkx as nx
@@ -14,8 +13,6 @@ import numpy as np
 import pandas as pd
 import re
 import time
-import xml
-import xml.sax
 from shapely import wkt
 from shapely.geometry import LineString
 from shapely.geometry import Point
@@ -805,75 +802,3 @@ def make_shp_filename(place_name):
     filename = '-'.join(name_pieces).lower().replace(' ','_')
     filename = re.sub('[^0-9a-zA-Z_-]+', '', filename)
     return filename
-
-
-
-def overpass_json_from_file(filename):
-    """
-    Read OSM XML from input filename and return Overpass-like JSON.
-
-    Parameters
-    ----------
-    filename : string
-        name of file containing OSM XML data
-
-    Returns
-    -------
-    OSMContentHandler object
-    """
-
-    _, ext = os.path.splitext(filename)
-
-    if ext == '.bz2':
-        # Use Python 2/3 compatible BZ2File()
-        opener = lambda fn: bz2.BZ2File(fn)
-    else:
-        # Assume an unrecognized file extension is just XML
-        opener = lambda fn: open(fn, mode='rb')
-
-    with opener(filename) as file:
-        handler = OSMContentHandler()
-        xml.sax.parse(file, handler)
-        return handler.object
-
-
-
-class OSMContentHandler(xml.sax.handler.ContentHandler):
-    """
-    SAX content handler for OSM XML.
-
-    Used to build an Overpass-like response JSON object in self.object. For format
-    notes, see http://wiki.openstreetmap.org/wiki/OSM_XML#OSM_XML_file_format_notes
-    and http://overpass-api.de/output_formats.html#json
-    """
-
-    def __init__(self):
-        self._element = None
-        self.object = {'elements': []}
-
-    def startElement(self, name, attrs):
-        if name == 'osm':
-            self.object.update({k: attrs[k] for k in attrs.keys()
-                                if k in ('version', 'generator')})
-
-        elif name in ('node', 'way'):
-            self._element = dict(type=name, tags={}, nodes=[], **attrs)
-            self._element.update({k: float(attrs[k]) for k in attrs.keys()
-                                  if k in ('lat', 'lon')})
-            self._element.update({k: int(attrs[k]) for k in attrs.keys()
-                                  if k in ('id', 'uid', 'version', 'changeset')})
-
-        elif name == 'tag':
-            self._element['tags'].update({attrs['k']: attrs['v']})
-
-        elif name == 'nd':
-            self._element['nodes'].append(int(attrs['ref']))
-
-        elif name == 'relation':
-            # Placeholder for future relation support.
-            # Look for nested members and tags.
-            pass
-
-    def endElement(self, name):
-        if name in ('node', 'way'):
-            self.object['elements'].append(self._element)
